% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchApp}
\alias{launchApp}
\title{Launch expenditR}
\usage{
launchApp(port = 3838)
}
\arguments{
\item{port}{The port, expenditR is running on (default: 3838)}
}
\value{
expenditR application
}
\description{
Launch expenditR
}
