% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCommunity.R
\name{moduleCommunityUI}
\alias{moduleCommunityUI}
\title{moduleCommunityUI}
\usage{
moduleCommunityUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleCommunityUI
}
