% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePrivate.R
\name{modulePrivateUI}
\alias{modulePrivateUI}
\title{modulePrivateUI}
\usage{
modulePrivateUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
modulePrivateUI
}
