% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSummary.R
\name{moduleSummaryUI}
\alias{moduleSummaryUI}
\title{moduleSummaryUI}
\usage{
moduleSummaryUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleSummaryUI
}
